Rebase nspire/main.c on newest unix/main.c
This is mostly based on reviewing the diff between the two and applying all
changes from unix/main.c. A few large blocks behind #ifdef's which aren't
applicable (like readline) are deleted to make the diff against the old
nspire/main.c a bit smaller.

This fixes the "Unknown NLR" issues which happened on exit with the latest
release (2014-10-27) and on every input with an updated micropython but old
main.c.

The new handling for sys.path and sys.argv seem to yield the same result than
the old one did when opening a .py.tns file.